backend/requirements.txt
fastapi == 0.65.1
httpx == 0.26.0
uvicorn == 0.14.0
motor == 3.1.1
python-Levenshtein == 0.23.0
python-dotenv == 1.0.0